+++ src/scopeActionCreators.js
@@ -1,6 +1,6 @@
-import { memoize } from './helpers/memoize';
+import { memoize } from './utils/memoize';
 
-import * as object from './helpers/object-shim';
+import * as object from './utils/object-shim';
 
 // Actions
 // =======
