+++ src/utils/object-shim.js
